+++ examples/deadline/All-In-AWS-Infrastructure-Basic/ts/bin/config.ts
@@ -6,6 +6,7 @@
 import 'source-map-support/register';
 import { UsageBasedLicense } from 'aws-rfdk/deadline';
 import { MongoDbSsplLicenseAcceptance } from 'aws-rfdk';
+import { AwsThinkboxEulaAcceptance } from 'aws-rfdk/deadline';
 
 /**
  * Configuration values for the sample app.
@@ -13,6 +14,21 @@ import { MongoDbSsplLicenseAcceptance } from 'aws-rfdk';
  * TODO: Fill these in with your own values.
  */
 class AppConfig {
+  /**
+   * Change this value to AwsThinkboxEulaAcceptance.USER_ACCEPTS_AWS_THINKBOX_EULA if you wish to accept the EULA for
+   * Deadline and proceed with Deadline deployment. Users must explicitly accept the AWS Thinkbox EULA before using the
+   * AWS Thinkbox Deadline container images.
+   *
+   * See https://www.awsthinkbox.com/end-user-license-agreement for the terms of the agreement.
+   */
+  public readonly acceptAwsThinkboxEula: AwsThinkboxEulaAcceptance = AwsThinkboxEulaAcceptance.USER_REJECTS_AWS_THINKBOX_EULA;
+
+  /**
+   * The version of Deadline to use on the render farm. Some examples of pinned version values are "10", "10.1", or
+   * "10.1.12"
+   * @default The latest available version of Deadline is used
+   */
+  public readonly deadlineVersion?: string;
 
   /**
    * A map of regions to Deadline Client Linux AMIs. As an example, the Linux Deadline 10.1.12.1 AMI ID from us-west-2
@@ -49,7 +65,6 @@ class AppConfig {
    * if you wish to accept the SSPL and proceed with MongoDB deployment.
    */
   public readonly acceptSsplLicense: MongoDbSsplLicenseAcceptance = MongoDbSsplLicenseAcceptance.USER_REJECTS_SSPL;
-
 }
 
 export const config = new AppConfig();